[common/meta] refactor: move SeqV common/meta{sled-store -> types}
