+++ common/meta/sled-store/Cargo.toml
@@ -10,6 +10,7 @@ edition = "2021"
 
 [dependencies]
 common-exception = {path = "../../exception"}
+common-meta-types = {path = "../types"}
 common-tracing = {path = "../../tracing"}
 
 anyhow = "1.0.44"
@@ -24,7 +25,6 @@ tempfile = "3.2.0"
 
 [dev-dependencies]
 common-base = {path = "../../base" }
-common-meta-types = {path = "../types"}
 
 pretty_assertions = "1.0"
