+++ common/meta/sled-store/src/testing/fake_key_spaces.rs
@@ -17,10 +17,10 @@ use common_meta_types::LogEntry;
 use common_meta_types::LogIndex;
 use common_meta_types::Node;
 use common_meta_types::NodeId;
+use common_meta_types::SeqV;
 
 use crate::testing::fake_state_machine_meta::StateMachineMetaKey;
 use crate::testing::fake_state_machine_meta::StateMachineMetaValue;
-use crate::SeqV;
 use crate::SledKeySpace;
 
 /// Types for raft log in SledTree
