+++ common/meta/types/Cargo.toml
@@ -9,7 +9,6 @@ edition = "2021"
 [dependencies]
 common-datavalues = {path = "../../datavalues"}
 common-exception = {path = "../../exception"}
-common-meta-sled-store = {path = "../sled-store"}
 
 serde = { version = "1.0", features = ["derive"] }
 serde_json = "1.0"
