+++ common/meta/types/src/lib.rs
@@ -19,12 +19,10 @@ pub use cluster::NodeInfo;
 pub use cluster::Slot;
 pub use cmd::Cmd;
 pub use commit_table_reply::UpsertTableOptionReply;
-pub use common_meta_sled_store::IntoSeqV;
-pub use common_meta_sled_store::KVMeta;
-pub use common_meta_sled_store::SeqV;
 pub use database_info::DatabaseInfo;
 pub use database_reply::CreateDatabaseReply;
 pub use errors::ConflictSeq;
+pub use kv::KVMeta;
 pub use kv_reply::GetKVActionReply;
 pub use kv_reply::MGetKVActionReply;
 pub use kv_reply::PrefixListReply;
@@ -41,6 +39,8 @@ pub use raft_types::LogIndex;
 pub use raft_types::NodeId;
 pub use raft_types::Term;
 pub use seq_num::SeqNum;
+pub use seq_value::IntoSeqV;
+pub use seq_value::SeqV;
 pub use table_info::TableIdent;
 pub use table_info::TableInfo;
 pub use table_info::TableMeta;
@@ -52,6 +52,7 @@ mod commit_table_reply;
 mod database_info;
 mod database_reply;
 mod errors;
+mod kv;
 mod kv_reply;
 mod log_entry;
 mod match_seq;
@@ -59,6 +60,7 @@ mod operation;
 mod raft_txid;
 mod raft_types;
 mod seq_num;
+mod seq_value;
 mod table_info;
 mod table_reply;
